+++ functions/public/Get-FunctionName.ps1
@@ -2,7 +2,12 @@ Function Get-FunctionName {
     [cmdletbinding()]
     [outputType("string","PSFunctionName")]
     Param(
-        [Parameter(Position = 0, Mandatory, HelpMessage = "Specify the .ps1 or .psm1 file with defined functions.")]
+        [Parameter(
+            Position = 0,
+            Mandatory,
+            ValueFromPipeline,
+            HelpMessage = "Specify the .ps1 or .psm1 file with defined functions."
+        )]
         [ValidateScript({
             If (Test-Path $_ ) {
                 $True