+++ functions/public/New-CommentHelp.ps1
@@ -1,5 +1,5 @@
 Function New-CommentHelp {
-    [cmdletbinding()]
+    [cmdletbinding(DefaultParameterSetName="ast")]
     [alias('nch')]
     [OutputType("string")]
     Param(
@@ -7,13 +7,16 @@ Function New-CommentHelp {
             Position = 0,
             Mandatory,
             ValueFromPipeline,
-            HelpMessage = "Specify ParamBlockAST object."
+            HelpMessage = "Specify ParamBlockAST object.",
+            ParameterSetName = "ast"
         )]
         [System.Management.Automation.Language.ParamBlockAst]$ParamBlock,
         [Parameter(HelpMessage = "Provide a short description. You can always edit this later.")]
         [string]$Synopsis = "<short description>",
         [Parameter(HelpMessage = "Provide a longer description. You can always edit this later.")]
-        [string]$Description = "<long description>"
+        [string]$Description = "<long description>",
+        [Parameter(HelpMessage = "Insert help template",ParameterSetName = "template")]
+        [switch]$TemplateOnly
     )
 
     Begin {
@@ -21,40 +24,43 @@ Function New-CommentHelp {
         Write-Verbose "Defining help opening"
         $h = [System.Collections.Generic.List[string]]::new()
         $h.Add("<#")
-        $h.Add("`t.Synopsis")
-        $h.Add("`t  $Synopsis")
-        $h.add("`t.Description")
-        $h.add("`t  $Description")
+        $h.Add("  .Synopsis")
+        $h.Add("    $Synopsis")
+        $h.Add("  .Description")
+        $h.Add("    $Description")
     }
 
     Process {
-        Write-Verbose "Processing a paramblock AST object"
-        foreach ($p in $ParamBlock.Parameters) {
-            $paramName = $p.name.variablepath.userpath
-            Write-Verbose "Adding parameter help for $paramname"
-            $h.Add("`t.Parameter $paramName")
-            $paramHelp = $p.Attributes.namedArguments.where({ $_.argumentname -eq 'helpmessage' })
-            if ($paramHelp) {
-                $h.add("`t  $($paramHelp.argument.value)")
+        if ($PSCmdlet.ParameterSetName -eq 'ast') {
+            Write-Verbose "Processing a paramblock AST object"
+            foreach ($p in $ParamBlock.Parameters) {
+                $paramName = $p.name.variablepath.userpath
+                Write-Verbose "Adding parameter help for $paramname"
+                $h.Add("  .Parameter $paramName")
+                $paramHelp = $p.Attributes.namedArguments.where({ $_.argumentname -eq 'helpmessage' })
+                if ($paramHelp) {
+                    $h.Add("    $($paramHelp.argument.value)")
+                }
+                else {
+                    $h.Add("    <enter a parameter description>")
+                }
             }
-            else {
-                $h.Add("`t  <enter a parameter description>")
-            }
-        }
+        } #if paramblock
+
     } #process
     End {
         Write-Verbose "Adding closing help content"
-        $h.add("`t.Example")
-        $h.Add("`t  PS C:\> $Name")
-        $h.Add("`t  <output and explanation>")
-        $h.Add("`t.Inputs")
-        $h.add("`t  <Inputs to this function (if any)>")
-        $h.Add("`t.Outputs")
-        $h.add("`t  <Output from this function (if any)>")
-        $h.Add("`t.Notes")
-        $h.Add("`t  <General notes>")
-        $h.Add("`t.Link")
-        $h.Add("`t  <enter a link reference>")
+        $h.Add("  .Example")
+        $h.Add("    PS C:\> $Name")
+        $h.Add("    <output and explanation>")
+        $h.Add("  .Inputs")
+        $h.Add("    <Inputs to this function (if any)>")
+        $h.Add("  .Outputs")
+        $h.Add("    <Output from this function (if any)>")
+        $h.Add("  .Notes")
+        $h.Add("    <General notes>")
+        $h.Add("  .Link")
+        $h.Add("    <enter a link reference>")
         $h.Add("#>")
         $h
         Write-Verbose "Ending $($MyInvocation.MyCommand)"