+++ samples/Demo-NewModuleFromFiles.ps1
@@ -1,0 +1,14 @@
+#this demo assumes you have the Platyps module and git installed
+
+$splat = @{
+    Description   = "Demo exported module"
+    Files         = "$PSScriptRoot\tools.psm1"
+    Layout        = "$PSScriptRoot\ModuleLayout.json"
+    NewModuleName = "PSFooExport"
+    ParentPath    = $env:temp
+    Verbose       = $True
+    CreateHelp    = $True
+    FunctionPath  = "functions\public"
+    InitializeGit = $true
+}
+New-ModuleFromFiles @splat