+++ samples/SampleScript.ps1
@@ -1,0 +1,7 @@
+﻿#requires -version 3.0
+
+#this is a sample script
+
+Write-Host "this is a sample script that doesn't do anything but write a random number" -ForegroundColor Yellow
+
+Get-Random -Minimum 1 -Maximum 1000